+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - [#1870] AS3 Adding new class - Target DoABC tag or position can be selected to prevent Error 1014
+- [#1871] Toogle buttons for disabling subsprite animation, display preview of sprites/frames
 
 ### Fixed
 - [#1869] Replace references now replaces all references, not just PlaceObject
@@ -2586,6 +2587,7 @@ All notable changes to this project will be documented in this file.
 [alpha 8]: https://github.com/jindrapetrik/jpexs-decompiler/compare/alpha7...alpha8
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
 [#1870]: https://www.free-decompiler.com/flash/issues/1870
+[#1871]: https://www.free-decompiler.com/flash/issues/1871
 [#1869]: https://www.free-decompiler.com/flash/issues/1869
 [#1872]: https://www.free-decompiler.com/flash/issues/1872
 [#1867]: https://www.free-decompiler.com/flash/issues/1867